my modifications for dcplaya from a long time ago

+++ kernel/fs/fs_ramdisk.c
@@ -628,7 +628,13 @@ int fs_ramdisk_detach(const char * fn, void ** obj, size_t * size) {
 }
 
 /* Initialize the file system */
+static int init; /* VP : added already initialized check */
 int fs_ramdisk_init() {
+        if (init)
+	  return 0;
+
+	init = 1;
+
 	/* Create an empty root dir */
 	rootdir = (rd_dir_t *)malloc(sizeof(rd_dir_t));
 	LIST_INIT(rootdir);
@@ -654,6 +660,12 @@ int fs_ramdisk_init() {
 /* De-init the file system */
 int fs_ramdisk_shutdown() {
 	rd_file_t *f1, *f2;
+
+	if (!init)
+	  return 0;
+	
+	init = 0;
+
 	/* For now assume there's only the root dir, since mkdir and
 	   rmdir aren't even implemented... */
 	f1 = LIST_FIRST(rootdir);
